+++ frontend/pages/procreg/[submission_id]/[slug].vue
@@ -2,7 +2,7 @@
 import { useQuery } from "@vue/apollo-composable";
 import { graphql } from "~/generated/gql";
 import type { GetFormQuery } from "~/generated/gql/graphql";
-import FormWrapper from "~/components/shared/FormWrapper.vue";
+import FormWrapper from "~/components/form/FormWrapper.vue";
 
 const GET_FORM = graphql(`
     query GetForm($submissionId: ID!) {
